chore: refactor to split up components in heroDiagram
