+++ docs/.vitepress/theme/components/landing/hero-diagram/SvgNode.vue
